requirements.txt
streamlit
mysql-connector-python
bcrypt
cryptography
pytesseract
pillow